requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
ffmpeg-python==0.2.0
PyNaCl==1.5.0